+++ test/list_test.exs
@@ -54,4 +54,22 @@ defmodule Moar.ListTest do
       assert Moar.List.to_sentence(["ant", "", "bat", nil, "cat"], &String.upcase/1) == "ANT, BAT, and CAT"
     end
   end
+
+  describe "unwrap!" do
+    test "returns the only term in the list" do
+      assert Moar.List.unwrap!(["a"]) == "a"
+    end
+
+    test "returns the term if it is not a list" do
+      assert Moar.List.unwrap!("a") == "a"
+    end
+
+    test "blows up if the list has more than one item" do
+      assert_raise FunctionClauseError, fn -> Moar.List.unwrap!(["a", "b"]) end
+    end
+
+    test "blows up if the list is empty" do
+      assert_raise FunctionClauseError, fn -> Moar.List.unwrap!([]) end
+    end
+  end
 end